rename todo hooks
