+++ src/components/pages/Todos/TodosListPage/index.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from "react";
-import { useTodosList } from "~/hooks/usecases/useTodosList";
+import { useTodosList } from "~/hooks/usecases/Todo/useTodosList";
 import Skeleton from "react-loading-skeleton";
 import { TodoColumnBoard } from "~/components/Todos/TodoColumnBoard";
